refactor map-overview component
