+++ src/app/map/map.component.scss
@@ -2,7 +2,7 @@
 
 @import '../../theming/typography';
 
-.map-overview {
+.map {
   height: 100%;
 
   map-info-window {
